Add definitions for Endpoint Companion and BOS descriptors. Add APIs for parsing the Endpoint Companion and BOS descriptors.
Nathan: Modified the original patch to fit the interface now already
provided by the FreeBSD implementation of libusb 1.0 and fit with
the rest of libusb.h.

Some discussion: The FreeBSD API is not optimal. Further discussion
is needed to determine how to update the API without breaking code
that used the existing API. This discussion should probably involve
the FreeBSD developers and could be handled in a 1.0.x release.

An example has been added to show how to use the new API.

Fixes #110